feat: add solution modules
